+++ src/components/Travel/Accommodations.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Link from './Link';
+import Link from '../Link';
 
 export default function Accommodations() {
   return (
